requirements.txt
pillow==8.4.0
pytesseract==0.3.8
gtts==2.2.3